Shortened the program
